Producto/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2
django-heroku
gunicorn